Consolidate integration tests, rename helpers, freeze dataclasses
Folds the 11 per-axis test files added in the previous commit into 5
focused modules: test_literalize_golden (six literalize → golden axes),
test_calls (default + variant call goldens), test_orphan_golden_files,
test_check_golden. Moves test_format_enumeration_properties to
tests/test_languages.py since it doesn't touch goldens. Renames
golden_assertions.py to check_golden.py (single function), folds
variable_form_wrapping into variant_cases. Switches the orphan test to
the cases_dir fixture, marks every dataclass frozen=True, and applies
@beartype to the helper functions that lacked it for consistency.
