+++ tests/integration/test_calls.py
@@ -1,5 +1,9 @@
-"""``literalize_call`` golden-file tests against each language's default
-spec.
+"""``literalize_call`` golden-file tests.
+
+Drives each call-case configuration against every supporting language,
+both for default specs and for non-default language variants (e.g.
+Rust's ``TAGGED_ENUM`` strategy on inputs the default ``ERROR``
+strategy rejects).
 """
 
 from pathlib import Path
@@ -8,6 +12,7 @@ import pytest
 from pytest_regressions.file_regression import FileRegressionFixture
 
 from .call_cases import CallCase, discover_call_cases, run_call_golden_case
+from .call_variant_cases import CallVariantCase, build_call_variant_cases
 from .language_specs import make_spec
 
 
@@ -40,6 +45,36 @@ def test_call_golden_file(
         config=config,
         spec=spec,
         golden_name=f"{lang_cls.__name__}_call",
+        cases_dir=cases_dir,
+        file_regression=file_regression,
+    )
+
+
+@pytest.mark.parametrize(
+    argnames="call_variant_case",
+    argvalues=build_call_variant_cases(),
+    ids=[
+        f"{c.config.case_dir_name}/{c.variant.name}"
+        for c in build_call_variant_cases()
+    ],
+)
+def test_call_variant_golden_file(
+    call_variant_case: CallVariantCase,
+    cases_dir: Path,
+    file_regression: FileRegressionFixture,
+) -> None:
+    """Test ``literalize_call`` for a non-default language spec.
+
+    Covers call inputs that the language's default
+    :attr:`Language.heterogeneous_strategy` rejects, which
+    :func:`test_call_golden_file` skips — in particular the
+    sibling-widening behavior of Rust's ``TAGGED_ENUM`` across
+    per-element call arguments.
+    """
+    run_call_golden_case(
+        config=call_variant_case.config,
+        spec=call_variant_case.variant.spec,
+        golden_name=f"{call_variant_case.variant.name}_call",
         cases_dir=cases_dir,
         file_regression=file_regression,
     )
